+++ src/main/java/com/controllers/EventController.java
@@ -8,7 +8,7 @@ import com.serviceLayer.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import java.util.HashSet;
 import java.util.Set;
@@ -21,30 +21,55 @@ public class EventController {
     @Autowired
     UserService userService;
 
-    @RequestMapping(value = "/addResponsibleUser",method = RequestMethod.POST)
-    public String setResponsibleUser(HttpSession session, @RequestParam("eventId") int eventId){
+    @RequestMapping(value = "/addResponsibleUser", method = RequestMethod.POST)
+    public String setResponsibleUser(HttpSession session, @RequestParam("eventId") int eventId) {
         int userId = (int) session.getAttribute("userId");
-        eventService.setResponsibleUser(userId,eventId);
+        eventService.setResponsibleUser(userId, eventId);
         return "redirect:/events/";
     }
 
     @RequestMapping("/eventsJson/")
-    public @ResponseBody
+    public
+    @ResponseBody
     Set<EventDTO> getEvents(HttpSession session) {
         Set<EventDTO> EventDTOs = new HashSet<>();
         User user = userService.getUser((int) session.getAttribute("userId"));
-        for (Event e: eventService.getListOfAllEvents()) {
-            EventDTOs.add(new EventDTO(e,user));
+        for (Event event : eventService.getListOfAllEvents()) {
+            EventDTOs.add(new EventDTO(event, user));
         }
         return EventDTOs;
     }
 
     @RequestMapping("/event_{id}")
-    public @ResponseBody
+    public
+    @ResponseBody
     EventDTO getEventById(@PathVariable("id") int eventId) {
         EventDTO EventDTO = new EventDTO(eventService.getEventById(eventId));
         return EventDTO;
     }
 
+    @RequestMapping(value = "/events", method = RequestMethod.POST)
+    public String events(HttpSession session, HttpServletRequest req) {
+        String email = req.getParameter("email");
+        session.setAttribute("backPage","/");
+        userService.saveUser(email);
 
+        session.setAttribute("userId", userService.getUserByEmail(req.getParameter("email")).getId());
+        return "events";
+    }
+
+    @RequestMapping(value = "/new_event", method = RequestMethod.POST)
+    public String newEvent(HttpServletRequest req) {
+        eventService.saveByRequest(req);
+        String ref = req.getHeader("Referer");
+        return "redirect:" + ref;
+    }
+
+    @RequestMapping(value = "/events")
+    public String events(HttpSession session) {
+
+        session.setAttribute("allEvents", eventService.getListOfAllEvents());
+        session.setAttribute("backPage", "/events");
+        return "events";
+    }
 }
