+++ src/main/java/com/controllers/OrderController.java
@@ -9,8 +9,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import java.util.List;
 
@@ -68,4 +70,10 @@ public class OrderController {
         return orderService.orderListOfUserByEvent(userId,eventId);
     }
 
+    @RequestMapping(value = "/add_to_order", method = RequestMethod.POST)
+    public String addToOrder(HttpServletRequest req, HttpSession session) {
+        orderService.saveByRequest(req, session);
+        String ref = req.getHeader("Referer");
+        return "redirect:" + ref;
+    }
 }
