+++ src/main/java/com/controllers/RestaurantsController.java
@@ -1,0 +1,51 @@
+package com.controllers;
+
+import com.DTOLayer.DTOEntity.RestaurantDTO;
+import com.model.Entity.Restaurant;
+import com.serviceLayer.service.RestaurantService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.*;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+import java.util.ArrayList;
+import java.util.List;
+
+@Controller
+public class RestaurantsController {
+    @Autowired
+    RestaurantService restaurantService;
+
+    @RequestMapping(value = "/new_restaurant", method = RequestMethod.POST)
+    public String newRestaurant(HttpServletRequest req , @RequestParam("eventId") int eventId) {
+        restaurantService.saveByRequest(req);
+        String ref = req.getHeader("Referer");
+        return "redirect:" + ref+"#/"+eventId;
+    }
+
+    @RequestMapping(value = "/events/event_{id}")
+    public String getAllRestaurants(HttpSession session, @PathVariable("id") String id) {
+        session.setAttribute("eventId", id);
+        session.setAttribute("allRestaurants", restaurantService.getListOfAllRestaurant());
+        session.setAttribute("backPage", "/events");
+        return "restaurants";
+    }
+
+    @RequestMapping(value = "/events/event_{event}/restaurant_{id}")
+    public String getAllItemsByRestaurant(HttpSession session, @PathVariable("event") int event, @PathVariable("id") int id) {
+        session.setAttribute("restaurant", restaurantService.getRestaurantById(id));
+        session.setAttribute("Menu", restaurantService.getRestaurantById(id).getItem());
+        session.setAttribute("backPage", "/events/event_" + event);
+        return "menu";
+    }
+
+    @RequestMapping("/restaurants")
+    public  @ResponseBody
+    List<RestaurantDTO> getRestaurants() {
+        List<RestaurantDTO> RestaurantDTOs = new ArrayList<RestaurantDTO>();
+        for (Restaurant restaurant: restaurantService.getListOfAllRestaurant()) {
+            RestaurantDTOs.add(new RestaurantDTO(restaurant));
+        }
+        return RestaurantDTOs;
+    }
+}
