+++ src/main/java/com/serviceLayer/implementation/OrderServiceImpl.java
@@ -8,7 +8,7 @@ import com.model.Entity.Item;
 import com.model.Entity.Order;
 import com.model.Entity.User;
 import com.serviceLayer.service.EventService;
-import com.serviceLayer.service.MenuService;
+import com.serviceLayer.service.ItemService;
 import com.serviceLayer.service.OrderService;
 import com.serviceLayer.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,7 +24,7 @@ public class OrderServiceImpl implements OrderService {
     @Autowired
     OrderDAO orderDAO;
     @Autowired
-    MenuService menuService;
+    ItemService itemService;
     @Autowired
     UserService userService;
 
@@ -39,7 +39,7 @@ public class OrderServiceImpl implements OrderService {
         int event_id = Integer.parseInt(req.getParameter("event_id"));
         Event event = eventService.getEventById(event_id);
         int item_id = Integer.parseInt(req.getParameter("item_id"));
-        Item item = menuService.getItemById(item_id);
+        Item item = itemService.getItemById(item_id);
         int user_id = (int) session.getAttribute("userId");
         User user = userService.getUser(user_id);
