+++ src/main/java/com/serviceLayer/service/ItemService.java
@@ -6,7 +6,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
 
-public interface MenuService {
+public interface ItemService {
     public void save(Item item);
 
     public void saveByRequest(HttpServletRequest req, HttpSession session);
