+++ src/main/webapp/WEB-INF/views/jsp/partials/Edit.jsp
@@ -1,10 +1,3 @@
-<%--
-  Created by IntelliJ IDEA.
-  User: employee
-  Date: 4/12/16
-  Time: 10:00 AM
-  To change this template use File | Settings | File Templates.
---%>
 <div ng-controller="eventController">
     {{event.name}} {{event.description}}
 </div>
