+++ src/main/webapp/WEB-INF/views/jsp/partials/comonOrderByEachUser.jsp
@@ -1,12 +1,9 @@
 <div ng-controller="commonOrderByEachUserCtrl" class="container">
-
     <a href="#/commonOrder/{{id}}">
         <button type="button" class="btn btn-success ">Common</button>
     </a>
-
     <div class="row">
         <div class="col-md-7">
-
             <div>
                 <h3>Common order by users</h3>
                 <div ng-repeat="orderDTO in array">
@@ -14,7 +11,6 @@
                         <tr>
                             {{orderDTO.userMail}}
                         </tr>
-
                         <tr>
                             <th>Restaurant</th>
                             <th>Name</th>
@@ -35,5 +31,4 @@
             </div>
         </div>
     </div>
-
 </div>
